+++ app/src/main/java/kr/co/lion/modigm/ui/login/email/EmailLoginFragment.kt
@@ -1,4 +1,4 @@
-package kr.co.lion.modigm.ui.login
+package kr.co.lion.modigm.ui.login.email
 
 import android.os.Bundle
 import android.text.Editable
@@ -13,6 +13,9 @@ import kr.co.lion.modigm.R
 import kr.co.lion.modigm.databinding.FragmentEmailLoginBinding
 import kr.co.lion.modigm.ui.VBBaseFragment
 import kr.co.lion.modigm.ui.join.JoinFragment
+import kr.co.lion.modigm.ui.login.CustomLoginErrorDialog
+import kr.co.lion.modigm.ui.login.FindEmailFragment
+import kr.co.lion.modigm.ui.login.FindPasswordFragment
 import kr.co.lion.modigm.ui.login.vm.LoginViewModel
 import kr.co.lion.modigm.ui.study.BottomNaviFragment
 import kr.co.lion.modigm.util.FragmentName
